requirements.txt
elasticsearch
jsonschema
python-dotenv
requests
streamlit
